Extract AbstractBeanPropertyDefinition and PropertyFilterDefinition
this comes from the FW vaadin/framework#10347
